pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "file2link-clean"
version = "0.1.0"
description = "Fresh FastAPI Telegram file streaming and download service"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "telethon",
  "jinja2",
  "python-dotenv",
]
